MVVM restructure
