+++ Phone/app/src/main/java/com/facilitation/phone/viewModel/DashboardViewModel.kt
@@ -1,13 +1,13 @@
-package com.facilitation.phone.ui.notifications
+package com.facilitation.phone.viewModel
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class NotificationsViewModel : ViewModel() {
+class DashboardViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is notifications Fragment"
+        value = "This is dashboard Fragment"
     }
     val text: LiveData<String> = _text
 }
