+++ Phone/app/src/main/java/com/facilitation/phone/viewModel/HomeViewModel.kt
@@ -1,13 +1,13 @@
-package com.facilitation.phone.ui.dashboard
+package com.facilitation.phone.viewModel
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class DashboardViewModel : ViewModel() {
+class HomeViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is dashboard Fragment"
+        value = "This is home Fragment"
     }
     val text: LiveData<String> = _text
 }
